# Feat: speed up the witness assignment of RLP/Tx circuits