The calldata parts of RLP and Tx circuits involve a lot of `is_zero_chip.assign()` which will call `F.invert()`. Note that for inverting a vector of fields we can use the `batch_invert()` method to amortize the expensive operation `F.invert()`(In a modern cpu, one invert took about 17000 ns. 

We can postpone the invert operations to prover by using the `Assigned<F>` type. This trick applies to several other circuits (such as state_circuit).

A few data here.  Before this pr, the tx_circuit alone needs about **695s** to finish (212s x 2 = 424s in synthesis).
```
··End:     SimpleFloorPlanner synthesize ...........................................212.253s // should x2 as we have two phases.
End:     Tx Circuit Proof generation with degree = 22 ..............................694.402s
```
After the pr, the tx_circuit needs **280s** to finish(5.6s x2 = 11.2s in synthesis).
```
··End:     SimpleFloorPlanner synthesize ...........................................5.576s
End:     Tx Circuit Proof generation with degree = 22 ..............................280.423s
```